requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
pydicom==2.4.4
Pillow==10.2.0
numpy==1.26.3
transformers==4.36.0
torch==2.1.2
accelerate==0.25.0
